graph
graph with sample data set
